fix: Handle plain strings in --sync-aliases
It's valid and not that uncommon to have an alias specified by a single
string. So far, cargo-run-bin panics with a very unhelpful message when
it counters that. This commit fixes that. While I was at it, I also
turned other nearby unwrap()s into error messages to avoid similarly
bad UX with other unexpected (likely invalid) config.toml contents.